Fix batch logic (#11)
* Fix truncated docstring

* Fix wrong variable name

* Use a separate method to decide if a batch should be flushed

No real benefit yet, but should it make it easier to add a more complex
logic (e.g. based on message size) in the future.

* Check if a batch should be flushed every time a task is added

Without this, this is only done when closing the batch. So if lots of tasks are
added, the internal cache will grow a lot, which is not the desired effect of batching.

* Split batch tests into smaller functions